2004-03-25  Mike Kestner  <mkestner@ximian.com>
	* gtk/Gtk.metadata : Widget.Events is a Gdk.EventMask, not int
	* sample/GtkDemo/DemoDrawingArea.cs : remove int casts
	* sample/Scribble.cs : remove int casts

svn path=/trunk/gtk-sharp/; revision=24601

+++ sample/GtkDemo/DemoDrawingArea.cs
@@ -84,10 +84,8 @@ namespace GtkDemo {
 			// Ask to receive events the drawing area doesn't normally
 			// subscribe to
 
-			drawingArea1.Events = (int)EventMask.LeaveNotifyMask |
-			                     (int)EventMask.ButtonPressMask |
-			                     (int)EventMask.PointerMotionMask |
-			                     (int)EventMask.PointerMotionHintMask;
+			drawingArea1.Events = EventMask.LeaveNotifyMask | EventMask.ButtonPressMask |
+			                      EventMask.PointerMotionMask | EventMask.PointerMotionHintMask;
 			
 			window.ShowAll ();
 		}
